move component to thier own folder
