+++ src/web/components/StarRating/StarRating.stories.tsx
