+++ src/web/components/StarRating/StarRating.tsx
