+++ src/web/components/StarRating/StarRatingPadding.tsx
